+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added Portuguese (BR) locale (`pt_br` in config.yml)
 * Locales with special characters like Russian or Korean will now work properly and won't be seen as `?` characters
 * Added configurable murderer thrown sword speed
+* Added 1.14 sign game states (thanks to montlikadani)
 
 ### 1.1.3 Release (01/08.06.2019)
 * Fixed boss bar when disabled could throw errors breaking the game